requirements.txt
pymupdf==1.23.6
ibm-generative-ai==0.5.0
ibm-watson-machine-learning==1.0.311
langchain==0.0.333
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.11
rouge==1.0.1
streamlit==1.24.1
pymilvus
huggingface_hub
chromadb==0.4.21
sentence-transformers==2.2.2